Refine admin navigation and add permission page
