+++ src/pages/index.tsx
@@ -88,3 +88,4 @@ export { default as VotePage } from "./VotePage";
 export { default as JoinVotePage } from "./JoinVotePage";
 export { default as MeetCreate } from "./MeetCreate";
 export { default as VoteManage } from "./VoteManage";
+export { default as AdminPermission } from "./AdminPermission";
